# feat: include app root node_modules folder