--- a/build/scons-xcode-project-build.js
+++ b/build/scons-xcode-project-build.js
@@ -29,7 +29,7 @@ async function generateIndexJSON(dirToTraverse) {
 			if (fs.existsSync(file)) {
 				if (fs.statSync(file).isDirectory()) {
 					walk(file);
-				} else if (/\.js(on)?$/.test(filename)) {
+				} else if (/\.(c?js|json)$/.test(filename)) { // TODO: Support mjs files!
 					index[file.replace(/\\/g, '/').replace(dirToTraverse + '/', 'Resources/')] = 1;
 				}
 			}