--- a/iphone/TitaniumKit/TitaniumKit/Sources/API/Bridge.m
+++ b/iphone/TitaniumKit/TitaniumKit/Sources/API/Bridge.m
@@ -41,7 +41,7 @@
       basename = [@"app" retain];
     } else {
       NSString *last = [[url path] lastPathComponent];
-      basename = [[last stringByReplacingOccurrencesOfString:@".js" withString:@""] retain];
+      basename = [[last stringByReplacingOccurrencesOfString:@".js" withString:@""] retain]; // FIXME: Handle cjs/mjs?
     }
   }
   return basename;