--- a/tests/.gitignore
+++ b/tests/.gitignore
@@ -1,1 +1,0 @@
-Resources/node_modules/