… build/lib/test/test.js
… cli/lib/tasks/copy-resources-task.js
… cli/lib/tasks/process-css-task.js
… iphone/cli/lib/gather.js
… tests/Resources/app.js
… tests/Resources/require.test.js
… tests/Resources/ti-mocha.js
… tests/Resources/ti.android.r.test.js
… tests/Resources/ti.blob.test.js
… tests/Resources/ti.filesystem.file.test.js
… tests/Resources/ti.geolocation.test.js
… tests/Resources/ti.media.audiorecorder.test.js
… tests/Resources/ti.platform.test.js
… tests/Resources/ti.ui.clipboard.test.js
… tests/Resources/ti.ui.listview.test.js
… tests/Resources/ti.ui.shortcut.test.js
… tests/Resources/ti.ui.tableview.test.js
… tests/Resources/ti.ui.test.js
… tests/Resources/ti.ui.view.test.js
… tests/Resources/ti.ui.webview.test.js
… tests/Resources/utilities/assertions.js
… tests/Resources/utilities/mocha-filter.js
… tests/Resources/utilities/utilities.js
… tests/babel.config.json
… tests/package.json