**JIRA:** https://jira.appcelerator.org/browse/TIMOB-26352

**Description:**
This is a **work in progress** PR to have our CLI build process include/handle a project's root `node_modules` folder into the app. 

@brentonhouse has a module that can be used via a CLI hook to straight-up copy the production dependencies into the app at https://www.npmjs.com/package/@titanium/module-copier
 This is very fast in gathering the production dependencies to narrow the set of directories necessary to copy but misses the mark on a couple things:
- You have to hook it up for your project via a hook manually
- It copies the files as-is, rather than doing any of the normal things we do on app resources (i.e. minify, transpile, source map, encrypt, etc)

This PR adds support for including production npm packages in the app's root `node_modules` folder for iOS app (so far).

The next steps here are to:
- [ ] Get this working on Android too
- [ ] Try to refactor android's build command to re-use as much common logic/code for the process as we can with the stuff I extracted/refactored out of iOS.

Another note here, that I did some informal timings against a local copy of @hansemannn app (an old fork) before/after this PR along with trying to naively add node_modules as a input folder without any of the production vs development dependency filtering. These timings are solely the gathering of the file listing to then process (i.e. gathering up all the JS/CSS/images/HTML etc that we later copy/minify/transpile/etc)

Gathering file listing w/ these changes (which include production `node_modules`)
------
38.6 ms
53.6 ms
43.3 ms

Without this PR (and w/o considering app `node_modules`)
------
62.5 ms
70 ms
66 ms

Without this PR (and including app `node_modules`)
-------
240 ms
260 ms
242 ms

So - pretty exciting that after the refactor and adding the `node_modules` traversal we're still faster than before.