@build 2020-01-24
<!--
  0 failure: 
  3 warning:  This PR has miles..., Commit <a href="h..., Commit <a href="h...
  2 messages
  
  DangerID: danger-id-Danger;
-->


<table>
  <thead>
    <tr>
      <th width="50"></th>
      <th width="100%" data-danger-table="true">Warnings</th>
    </tr>
  </thead>
  <tbody><tr>
      <td>:warning:</td>
      <td>This PR has milestone set to 9.1.0, but the version defined in package.json is 9.0.0
Please either:
- Update the milestone on the PR
- Update the version in package.json
- Hold the PR to be merged later after a release and version bump on this branch</td>
    </tr>
  
<tr>
      <td>:warning:</td>
      <td>

  Commit <a href="https://api.github.com/repos/appcelerator/titanium_mobile/commits/31fc0b234ff86036cc015818b3da1ce8ad57a10f">31fc0b234ff86036cc015818b3da1ce8ad57a10f</a> has a message "feat: work in progress on copying node_modules during cli build into user app" giving 1 errors:
- header must not be longer than 72 characters, current length is 77
  </td>
    </tr>
  
<tr>
      <td>:warning:</td>
      <td>

  Commit <a href="https://api.github.com/repos/appcelerator/titanium_mobile/commits/0b301e4ef37f8517286371f2ef98c7d86ebfd8c8">0b301e4ef37f8517286371f2ef98c7d86ebfd8c8</a> has a message "feat: work in progress on copying node_modules during cli build into user app" giving 1 errors:
- header must not be longer than 72 characters, current length is 77
  </td>
    </tr>
  </tbody>
</table>


<table>
  <thead>
    <tr>
      <th width="50"></th>
      <th width="100%" data-danger-table="true">Messages</th>
    </tr>
  </thead>
  <tbody><tr>
      <td>:book:</td>
      <td>

  :white_check_mark: All tests are passing
Nice one! All 3 tests are passing.
  </td>
    </tr>
  
<tr>
      <td>:book:</td>
      <td>

  :rotating_light: This PR has one or more commits with warnings/errors for commit messages not matching our configuration. You may want to squash merge this PR and edit the message to match our conventions, or ask the original developer to modify their history.
  </td>
    </tr>
  </tbody>
</table>


<p align="right">
  Generated by :no_entry_sign: <a href="https://danger.systems/js">dangerJS</a> against 400a08283e3fc521b9ad5c93c646c05d3ef6b6da
</p>