@brentonhouse 2020-01-24
@sgtcoolguy  - A couple of things to note.  I looked through the code and might just be missing them.

- Excluding native modules from being copied (https://github.com/brentonhouse/titanium-module-copier/blob/1521d6421295c7fbc21244d12550d1e6e13392d3/index.js#L126-L130)
- Exclude anything with `titanium.ignore` from being copied . (https://github.com/brentonhouse/titanium-module-copier/blob/1521d6421295c7fbc21244d12550d1e6e13392d3/index.js#L122-L124)
- Exclude anything with titanium.type === 'widget' .  (https://github.com/brentonhouse/titanium-module-copier/blob/1521d6421295c7fbc21244d12550d1e6e13392d3/index.js#L131-L142)
- Including peerDependencies . (https://github.com/brentonhouse/titanium-module-copier/blob/1521d6421295c7fbc21244d12550d1e6e13392d3/index.js#L117-L119)