@brentonhouse 2020-01-24
Also, I took a look at the node_modules in the build directory and it looks like everything copied using `@titanium/module-copier` gets minified, source mapped, etc.