@sgtcoolguy 2020-01-27
@brentonhouse The issue with the `@titanium/module-copier` w/ hook approach right now is that it leads to extra file I/O (and you have to make sure you're hooking at the right point in the build cycle so that it does get processed). In all likelihood, I think it makes sense to do a PR on it to export more than just the one sync copying method so that you can use the library to give you the directory set back (which is what I did in the local copy here) and then the SDK CLI build can use that to get the input set.