@sgtcoolguy 2020-02-07
Sounds like #11346 may make this moot (at least going forward if people transition to using Webpack)