@brentonhouse 2020-02-07
> Sounds like #11346 may make this moot (at least going forward if people transition to using Webpack)

While webpack is cool, there are a lot of breaking changes in it and you can probably expect a good percentage of enterprises and devs that won't be able to or choose not to use webpack.