@sgtcoolguy 2020-04-14
This PR has a change to force it to use a specific branch of the test suite that verifies it's behavior. The change to the Jenkinsfile will need to be reverted/removed before merge. Additionally, after this is merged, the change to the test suite should likely be merged over on master to ensure we continue to test root-level package.json usage.