@sgtcoolguy 2021-01-25
Open questions/tasks:
  - handling of alloy widgets
  - turn off most/some of this processing for webpack projects?
  - Should `devDependencies` sometimes be included (i.e. if deploy type is test/development)?
    - Specifically, what if a developer wants to use something like mocha/should/jest/chai to test their app via sim/device, but not include in production builds?